Restore support for improperly encoded strings
Since the `json` gem was initially written for Ruby 1.8
before strings had encoding, it used to do its own UTF-8
validation direction on bytes and never really considered
the string declared encoding. So passing a ASCII-8BIT string
worked as long as the bytes were valid UTF-8

We may want to deprecate this, but we should emit warnings first.